refactor(neuron-ui): complement hook deps in several components, remove useless code, rename some method names
