--- packages/neuron-ui/src/components/Addresses/Address.ts
@@ -1,7 +1,0 @@
-export const Change = 'Change'
-export const Receiving = 'Receiving'
-
-export interface Address {
-  type: string
-  address: string
-}
