+++ packages/neuron-ui/src/containers/MainContent/actionCreators/wallets.ts
@@ -1,4 +1,7 @@
-import { initState, MainActions } from '../reducer'
+import {
+  //  initState,
+  MainActions,
+} from '../reducer'
 import { walletsCall } from '../../../services/UILayer'
 
 export default {
@@ -34,30 +37,8 @@ export default {
       payload: id,
     }
   },
-  importMnemonic: (params: { name: string; password: string; mnemonic: string }) => {
-    walletsCall.importMnemonic(params)
-    return {
-      type: MainActions.Wallet,
-    }
-  },
-  importKeystore: (params: { name: string; keystore: string; password: string }) => {
-    walletsCall.importKeystore(params)
-    return {
-      type: MainActions.Wallet,
-    }
-  },
   updateWallet: (params: { id: string; password: string; newPassword?: string; name?: string }) => {
     walletsCall.update(params)
-    return {
-      type: MainActions.Wallet,
-    }
-  },
-  importWallet: (isKeystore: boolean, params: typeof initState.tempWallet) => {
-    if (isKeystore) {
-      walletsCall.importKeystore(params)
-    } else {
-      walletsCall.importMnemonic(params)
-    }
     return {
       type: MainActions.Wallet,
     }
